# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Online Voting System with Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  message?: string;
}

export type ElectionStatus =
  (typeof ElectionStatus)[keyof typeof ElectionStatus];

export const ElectionStatus = {
  upcoming: "upcoming",
  active: "active",
  closed: "closed",
} as const;

export interface Election {
  id: number;
  title: string;
  description?: string;
  status: ElectionStatus;
  startDate: string;
  endDate: string;
  totalVotes: number;
  /** AI-computed fraud risk score 0-100 */
  fraudScore: number;
  createdAt: string;
}

export interface Candidate {
  id: number;
  electionId: number;
  name: string;
  party?: string;
  bio?: string;
  voteCount: number;
  percentage: number;
}

export type ElectionDetail = Election & {
  candidates: Candidate[];
};

export interface CreateElectionRequest {
  title: string;
  description?: string;
  startDate: string;
  endDate: string;
}

export type UpdateElectionRequestStatus =
  (typeof UpdateElectionRequestStatus)[keyof typeof UpdateElectionRequestStatus];

export const UpdateElectionRequestStatus = {
  upcoming: "upcoming",
  active: "active",
  closed: "closed",
} as const;

export interface UpdateElectionRequest {
  title?: string;
  description?: string;
  status?: UpdateElectionRequestStatus;
  startDate?: string;
  endDate?: string;
}

export interface CreateCandidateRequest {
  name: string;
  party?: string;
  bio?: string;
}

export type CastVoteRequestMetadata = { [key: string]: unknown };

export interface CastVoteRequest {
  electionId: number;
  candidateId: number;
  /** Anonymous voter identifier */
  voterId: string;
  voterIp?: string;
  metadata?: CastVoteRequestMetadata;
}

export interface VoteResponse {
  success: boolean;
  voteId: number;
  fraudFlag: boolean;
  message?: string;
}

export interface VoteStatus {
  hasVoted: boolean;
  candidateId?: number;
}

export interface ElectionResults {
  election: Election;
  candidates: Candidate[];
  totalVotes: number;
  turnoutPercentage: number;
  fraudAlertCount: number;
  winner?: Candidate;
}

export type FraudAlertType =
  (typeof FraudAlertType)[keyof typeof FraudAlertType];

export const FraudAlertType = {
  duplicate_ip: "duplicate_ip",
  rapid_voting: "rapid_voting",
  unusual_pattern: "unusual_pattern",
  bot_behavior: "bot_behavior",
  coordinated_attack: "coordinated_attack",
  suspicious_timing: "suspicious_timing",
} as const;

export type FraudAlertSeverity =
  (typeof FraudAlertSeverity)[keyof typeof FraudAlertSeverity];

export const FraudAlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface FraudAlert {
  id: number;
  electionId: number;
  electionTitle?: string;
  type: FraudAlertType;
  severity: FraudAlertSeverity;
  description: string;
  affectedVotes: number;
  voterId?: string;
  voterIp?: string;
  resolved: boolean;
  resolvedAt?: string;
  resolution?: string;
  createdAt: string;
}

export interface ResolveFraudAlertRequest {
  resolution: string;
  removeVotes?: boolean;
}

export type FraudAnalysisResultRiskLevel =
  (typeof FraudAnalysisResultRiskLevel)[keyof typeof FraudAnalysisResultRiskLevel];

export const FraudAnalysisResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type FraudAnalysisResultFindingsItem = {
  type: string;
  description: string;
  affectedVotes: number;
  severity: string;
};

export interface FraudAnalysisResult {
  electionId: number;
  analysisTimestamp: string;
  overallRiskScore: number;
  riskLevel: FraudAnalysisResultRiskLevel;
  findings: FraudAnalysisResultFindingsItem[];
  alertsCreated: number;
  recommendation: string;
}

export type DashboardSummaryElectionsByStatus = {
  upcoming: number;
  active: number;
  closed: number;
};

export interface DashboardSummary {
  totalElections: number;
  activeElections: number;
  totalVotesCast: number;
  fraudAlertsOpen: number;
  fraudAlertsResolved: number;
  electionsByStatus: DashboardSummaryElectionsByStatus;
  recentElections: Election[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  vote_cast: "vote_cast",
  fraud_detected: "fraud_detected",
  election_started: "election_started",
  election_closed: "election_closed",
  alert_resolved: "alert_resolved",
} as const;

export type ActivityItemSeverity =
  (typeof ActivityItemSeverity)[keyof typeof ActivityItemSeverity];

export const ActivityItemSeverity = {
  info: "info",
  warning: "warning",
  danger: "danger",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  electionId: number;
  electionTitle: string;
  severity: ActivityItemSeverity;
  timestamp: string;
}

export type FraudStatsAlertsByTypeItem = {
  type: string;
  count: number;
};

export type FraudStatsAlertsBySeverity = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export type FraudStatsTopRiskyElectionsItem = {
  electionId: number;
  electionTitle: string;
  fraudScore: number;
  alertCount: number;
};

export interface FraudStats {
  totalAlerts: number;
  alertsByType: FraudStatsAlertsByTypeItem[];
  alertsBySeverity: FraudStatsAlertsBySeverity;
  resolutionRate: number;
  topRiskyElections: FraudStatsTopRiskyElectionsItem[];
}

export type ListElectionsParams = {
  status?: ListElectionsStatus;
};

export type ListElectionsStatus =
  (typeof ListElectionsStatus)[keyof typeof ListElectionsStatus];

export const ListElectionsStatus = {
  active: "active",
  closed: "closed",
  upcoming: "upcoming",
} as const;

export type CheckVoteStatusParams = {
  electionId: number;
  voterId: string;
};

export type ListFraudAlertsParams = {
  electionId?: number;
  severity?: ListFraudAlertsSeverity;
  resolved?: boolean;
};

export type ListFraudAlertsSeverity =
  (typeof ListFraudAlertsSeverity)[keyof typeof ListFraudAlertsSeverity];

export const ListFraudAlertsSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};