# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-Powered Online Voting System with Fraud Detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /elections:
    get:
      operationId: listElections
      tags: [elections]
      summary: List all elections
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, closed, upcoming]
      responses:
        "200":
          description: List of elections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Election"
    post:
      operationId: createElection
      tags: [elections]
      summary: Create a new election
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateElectionRequest"
      responses:
        "201":
          description: Election created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{id}:
    get:
      operationId: getElection
      tags: [elections]
      summary: Get a single election with candidates
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Election details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateElection
      tags: [elections]
      summary: Update election status or details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateElectionRequest"
      responses:
        "200":
          description: Updated election
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{id}/results:
    get:
      operationId: getElectionResults
      tags: [elections]
      summary: Get vote tallies for an election
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Election results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionResults"
  /elections/{id}/candidates:
    post:
      operationId: addCandidate
      tags: [candidates]
      summary: Add a candidate to an election
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "201":
          description: Candidate added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
  /fraud/alerts:
    get:
      operationId: listFraudAlerts
      tags: [fraud]
      summary: List fraud alerts
      parameters:
        - in: query
          name: electionId
          schema:
            type: integer
        - in: query
          name: severity
          schema:
            type: string
            enum: [low, medium, high, critical]
        - in: query
          name: resolved
          schema:
            type: boolean
      responses:
        "200":
          description: List of fraud alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudAlert"
  /fraud/alerts/{id}/resolve:
    patch:
      operationId: resolveFraudAlert
      tags: [fraud]
      summary: Mark a fraud alert as resolved
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveFraudAlertRequest"
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudAlert"
  /fraud/analyze/{electionId}:
    post:
      operationId: analyzeElectionFraud
      tags: [fraud]
      summary: Trigger AI fraud analysis for an election
      parameters:
        - in: path
          name: electionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fraud analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudAnalysisResult"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    Election:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [upcoming, active, closed]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalVotes:
          type: integer
        fraudScore:
          type: number
          description: AI-computed fraud risk score 0-100
        createdAt:
          type: string
          format: date-time
      required: [id, title, status, startDate, endDate, totalVotes, fraudScore, createdAt]
    ElectionDetail:
      allOf:
        - $ref: "#/components/schemas/Election"
        - type: object
          properties:
            candidates:
              type: array
              items:
                $ref: "#/components/schemas/Candidate"
          required: [candidates]
    CreateElectionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [title, startDate, endDate]
    UpdateElectionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [upcoming, active, closed]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Candidate:
      type: object
      properties:
        id:
          type: integer
        electionId:
          type: integer
        name:
          type: string
        party:
          type: string
        bio:
          type: string
        voteCount:
          type: integer
        percentage:
          type: number
      required: [id, electionId, name, voteCount, percentage]
    CreateCandidateRequest:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        bio:
          type: string
      required: [name]
...
    ElectionResults:
      type: object
      properties:
        election:
          $ref: "#/components/schemas/Election"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        totalVotes:
          type: integer
        turnoutPercentage:
          type: number
        fraudAlertCount:
          type: integer
        winner:
          $ref: "#/components/schemas/Candidate"
      required: [election, candidates, totalVotes, turnoutPercentage, fraudAlertCount]
    FraudAlert:
      type: object
      properties:
        id:
          type: integer
        electionId:
          type: integer
        electionTitle:
          type: string
        type:
          type: string
          enum: [duplicate_ip, rapid_voting, unusual_pattern, bot_behavior, coordinated_attack, suspicious_timing]
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        affectedVotes:
          type: integer
        voterId:
          type: string
        voterIp:
          type: string
        resolved:
          type: boolean
        resolvedAt:
          type: string
          format: date-time
        resolution:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, electionId, type, severity, description, affectedVotes, resolved, createdAt]
    ResolveFraudAlertRequest:
      type: object
      properties:
        resolution:
          type: string
        removeVotes:
          type: boolean
      required: [resolution]
    FraudAnalysisResult:
      type: object
      properties:
        electionId:
          type: integer
        analysisTimestamp:
          type: string
          format: date-time
        overallRiskScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        findings:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
              affectedVotes:
                type: integer
              severity:
                type: string
            required: [type, description, affectedVotes, severity]
        alertsCreated:
          type: integer
        recommendation:
          type: string
      required: [electionId, analysisTimestamp, overallRiskScore, riskLevel, findings, alertsCreated, recommendation]
...