# lib/api-zod/src/generated/types/createElectionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Online Voting System with Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateElectionRequest {
  title: string;
  description?: string;
  startDate: Date;
  endDate: Date;
}