# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Online Voting System with Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
  message?: string;
}