# lib/api-zod/src/generated/types/updateElectionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Online Voting System with Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateElectionRequestStatus } from "./updateElectionRequestStatus";

export interface UpdateElectionRequest {
  title?: string;
  description?: string;
  status?: UpdateElectionRequestStatus;
  startDate?: Date;
  endDate?: Date;
}